added accept support for proxy query
allow wdqs-proxy to be queried using 'accept' as a GET parameter which will be converted to a POST parameter when making to request to Blazegraph.
this allows external tools like OpenRefine to load the data directly from a query in TSV instead of the JSON format.
only the valid accept types for TSV (text/tab-separated-values) and JSON (application/json) are supported, everything else results to the default XML. Those are the only types supported by Blazegraph.